[cli] Fix an error that caused the server connection to close
This commit fixes an issue where the connection to the server was
unexpectedly closed. It looks like the server did not keep the
session alive, causing it to close. As a result, when a new request
is made, it fails.

Now, when there is a connection issue with the GrimoireLab server,
the system attempts to reconnect and retry.

This change resolves the end-to-end tests running in GitHub Actions.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>